Change AI-based flow to state machine
