+++ app/backend/approaches/appresources.py
@@ -1,0 +1,37 @@
+import os
+import json
+import logging
+from typing import Optional
+
+import openai
+from azure.search.documents.aio import SearchClient
+
+from core.modelhelper import get_token_limit
+
+use_RAG = False
+
+class AppResources:
+    def __init__(
+        self,
+        search_client: SearchClient,
+        openai_host: str,
+        chatgpt_deployment: Optional[str],  # Not needed for non-Azure OpenAI
+        chatgpt_model: str,
+        embedding_deployment: Optional[str],  # Not needed for non-Azure OpenAI or for retrieval_mode="text"
+        embedding_model: str,
+        sourcepage_field: str,
+        content_field: str,
+        query_language: str,
+        query_speller: str,
+    ):
+        self.search_client = search_client
+        self.openai_host = openai_host
+        self.chatgpt_deployment = chatgpt_deployment
+        self.chatgpt_model = chatgpt_model
+        self.embedding_deployment = embedding_deployment
+        self.embedding_model = embedding_model
+        self.sourcepage_field = sourcepage_field
+        self.content_field = content_field
+        self.query_language = query_language
+        self.query_speller = query_speller
+        self.chatgpt_token_limit = get_token_limit(chatgpt_model)
